✨ Feature: 키보드 작업 되돌리기
